requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
bleach==6.1.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==5.1.1
django-debug-toolbar==4.4.6
django-rest-framework==0.1.0
django-summernote==0.8.20.0
django_ajax==0.3.0
djangorestframework==3.15.2
h11==0.14.0
hiredis==3.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
outcome==1.3.0.post0
Pillow==9.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
PySocks==1.7.1
redis==5.1.1
selenium==4.25.0
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
trio==0.26.2
trio-websocket==0.11.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
zope.interface==7.0.3
